change work flow to download deb packages

+++ cloudstack/big_patch.py
@@ -1,12 +1,29 @@
 # Copyright 2014 Big Switch Networks, Inc.
 # All Rights Reserved.
 #
-# This script requires sshpass python-yaml, python-pip,
-# python-dev and concurrent.futures on the patch node.
-# It also requires ssh on all nodes, i.e.,
-# sudo apt-get install -y sshpass python-yaml python-pip python-dev (on patch node)
-# sudo pip install futures subprocess32 (on patch node)
+# This script is used to set up cloud stack management node
+# and compute nodes with Big Cloud Fabric. The requirements are:
+# BCF: 2.0.1 or higher
+# installation node: ubuntu 12.04
+# management node: ubuntu 12.04
+# compute node: ubuntu 12.04 or xenserver 6.2
+# 
+# To prepare installation, on installation node, please download
+# cloudstack-common_4.5.0-snapshot_all.deb,
+# cloudstack-management_4.5.0-snapshot_all.deb,
+# cloudstack-agent_4.5.0-snapshot_all.deb
+# and put them under the same directory as this script.
+#
+# On installation node, run
+# sudo apt-get install -y sshpass python-yaml python-pip python-dev
+# sudo pip install futures subprocess32
+#
+# On all compute nodes, make sure ssh is installed
 # sudo apt-get install -y ssh (on all nodes)
+#
+# On installation node, edit example.yaml to reflect the physical setup, then
+# sudo python ./big_patch.py -c example.yaml
+
 
 '''
 # Following is an example data structure 
@@ -301,23 +318,8 @@ exec {"install maven3":
     returns => [0, 100],
 }
 
-exec {"wget cloudstack common":
-    path    => "/bin:/usr/bin:/usr/sbin",
-    command => "wget $cs_url/$cs_common -O /home/$user/bcf/$cs_common",
-    creates => "/home/$user/bcf/$cs_common",
-    timeout => 1200,
-}
-
-exec {"wget cloudstack management":
-    path    => "/bin:/usr/bin:/usr/sbin",
-    command => "wget $cs_url/$cs_mgmt -O /home/$user/bcf/$cs_mgmt",
-    creates => "/home/$user/bcf/$cs_mgmt",
-    timeout => 1200,
-}
-
 exec {"dpkg common":
-    require => [Exec['wget cloudstack common'],
-                Exec["install maven3"],
+    require => [Exec["install maven3"],
                 Exec['export nfs'],
                 Package['tomcat6'],
                 Package['jsvc'],
@@ -344,9 +346,7 @@ exec {"dpkg common":
 }
 
 exec {"dpkg management":
-    require => [Exec['wget cloudstack common'],
-                Exec['wget cloudstack management'],
-                Exec['dpkg common']],
+    require => Exec['dpkg common'],
     user    => root,
     path    => "/bin:/usr/bin:/usr/sbin:/sbin",
     command => "dpkg -i /home/$user/bcf/$cs_mgmt",
@@ -513,23 +513,8 @@ service {"libvirt-bin":
                 Service['dbus']],
 }
 
-exec {"wget cloudstack common":
-    path    => "/bin:/usr/bin:/usr/sbin",
-    command => "wget $cs_url/$cs_common -O /home/$user/bcf/$cs_common",
-    creates => "/home/$user/bcf/$cs_common",
-    timeout => 1200,
-}
-
-exec {"wget cloudstack agent":
-    path    => "/bin:/usr/bin:/usr/sbin",
-    command => "wget $cs_url/$cs_agent -O /home/$user/bcf/$cs_agent",
-    creates => "/home/$user/bcf/$cs_agent",
-    timeout => 1200,
-}
-
 exec {"dpkg common":
-    require => [Exec['wget cloudstack common'],
-                Package['ethtool'],
+    require => [Package['ethtool'],
                 Package['qemu-kvm'],
                 Package['ubuntu-vm-builder'],
                 Package['openjdk-7-jre'],
@@ -548,9 +533,7 @@ exec {"dpkg common":
 }
 
 exec {"dpkg agent":
-    require => [Exec['wget cloudstack common'],
-                Exec['wget cloudstack agent'],
-                Exec['dpkg common']],
+    require => Exec['dpkg common'],
     user    => root,
     path    => "/bin:/usr/bin:/usr/sbin:/sbin",
     command => "dpkg -i /home/$user/bcf/$cs_agent",
@@ -824,6 +807,8 @@ else
     echo "route del default" >> /etc/rc.local
     echo "route add default gw ${pxe_gw}" >> /etc/rc.local
 
+    # experimental, we may not need this at all
+    /opt/xensource/bin/xe-toolstack-restart
 fi
 '''
 
@@ -884,6 +869,9 @@ xe pool-join master-address=${master_address} master-username=${master_username}
 echo "sleep 60" >> /etc/rc.local
 echo "route del default" >> /etc/rc.local
 echo "route add default gw ${pxe_gw}" >> /etc/rc.local
+
+# experimental, we may not need this at all
+/opt/xensource/bin/xe-toolstack-restart
 '''
 
 XEN_IP_ASSIGNMENT=r'''
@@ -927,6 +915,7 @@ while [[ ${count} > 0 ]]; do
 done
 
 # configure bond interface ip
+ip_array=()
 bond_count=${#bond_inets[@]}
 slave_count=${#slave_name_labels[@]}
 for (( i=0; i<${slave_count}; i++ )); do
@@ -951,8 +940,26 @@ for (( i=0; i<${slave_count}; i++ )); do
         gateway=${bond_gateways[$k]}
         pif_uuid=$(xe pif-list host-name-label=${slave_name_label} device-name='' VLAN=${vlan} | grep -w uuid | grep -v network | awk '{print $NF}')
         xe pif-reconfigure-ip uuid=${pif_uuid} mode=${inet} IP=${ip} netmask=${mask} gateway=${gateway}
+        ip_array=("${ip_array[@]}" "${ip}")
         ping ${gateway} -c3
     done
+done
+
+ip_count=${#ip_array[@]}
+for (( i=0; i<${ip_count}; i++ )); do
+    ip=${ip_array[$i]}
+    count_down=30
+    while [[ ${count_down} > 0 ]]; do
+        ping ${ip} -c1
+        if [[ $? == 0 ]]; then
+            break
+        fi
+        let count_down-=1
+        sleep 1
+    done
+    if [[ ${count_down}==0 ]]; then
+        echo "Failed to assign IP:" "${ip}"
+    fi
 done
 '''
 
@@ -995,9 +1002,16 @@ if [[ ("%(role)s" == "management") || ("%(hypervisor)s" == "kvm") ]]; then
     sshpass -p %(pwd)s scp /tmp/%(hostname)s.intf %(user)s@%(hostname)s:/home/%(user)s/bcf/%(role)s.intf >> %(log)s 2>&1
     echo -e "Copy %(role)s.pp to node %(hostname)s\n"
     sshpass -p %(pwd)s scp /tmp/%(hostname)s.pp %(user)s@%(hostname)s:/home/%(user)s/bcf/%(role)s.pp >> %(log)s 2>&1
+    echo -e "Copy %(CS_COMMON)s to node %(hostname)s\n"
+    sshpass -p %(pwd)s scp /tmp/%(CS_COMMON)s %(user)s@%(hostname)s:/home/%(user)s/bcf/%(CS_COMMON)s >> %(log)s 2>&1
     if [ -f /tmp/%(hostname)s.db.sh ]; then
         echo -e "Copy db.sh to node %(hostname)s\n"
         sshpass -p %(pwd)s scp /tmp/%(hostname)s.db.sh %(user)s@%(hostname)s:/home/%(user)s/bcf/db.sh >> %(log)s 2>&1
+        echo -e "Copy %(CS_MGMT)s to node %(hostname)s\n"
+        sshpass -p %(pwd)s scp /tmp/%(CS_MGMT)s %(user)s@%(hostname)s:/home/%(user)s/bcf/%(CS_MGMT)s >> %(log)s 2>&1
+    else
+        echo -e "Copy %(CS_AGENT)s to node %(hostname)s\n"
+        sshpass -p %(pwd)s scp /tmp/%(CS_AGENT)s %(user)s@%(hostname)s:/home/%(user)s/bcf/%(CS_AGENT)s >> %(log)s 2>&1
     fi
     echo -e "Copy %(role)s.sh to node %(hostname)s\n"
     sshpass -p %(pwd)s scp /tmp/%(hostname)s.remote.sh %(user)s@%(hostname)s:/home/%(user)s/bcf/%(role)s.sh >> %(log)s 2>&1
@@ -1409,7 +1423,10 @@ def generate_command_for_node(node):
                                 'role'       : node.role,
                                 'pool'       : node.xenserver_pool,
                                 'log'        : LOG_FILENAME,
-                                'hypervisor' : HYPERVISOR})
+                                'hypervisor' : HYPERVISOR,
+                                'CS_COMMON'  : CS_COMMON,
+                                'CS_MGMT'    : CS_MGMT,
+                                'CS_AGENT'   : CS_AGENT})
         node_local_bash.close()
 
     if HYPERVISOR == "xen":
@@ -1510,6 +1527,26 @@ def worker_reboot_slave():
 
 def deploy_to_all(config):
     # install sshpass
+    safe_print("Prepare cloud stack packages\n")
+    run_command_on_local(
+        'sudo mkdir -p /tmp;'
+        'sudo rm /tmp/*.deb;'
+        'sudo cp %(CS_COMMON)s /tmp/;'
+        'sudo cp %(CS_MGMT)s /tmp/;'
+        'sudo cp %(CS_AGENT)s /tmp/' %
+       {'CS_COMMON' : CS_COMMON,
+        'CS_MGMT'   : CS_MGMT,
+        'CS_AGENT'  : CS_AGENT})
+    if not os.path.isfile("/tmp/%s" % CS_COMMON):
+       safe_print("%s is missing\n" % CS_COMMON)
+       return
+    if not os.path.isfile("/tmp/%s" % CS_MGMT):
+       safe_print("%s is missing\n" % CS_MGMT)
+       return
+    if not os.path.isfile("/tmp/%s" % CS_AGENT):
+       safe_print("%s is missing\n" % CS_AGENT)
+       return
+
     safe_print("Installing sshpass to local node...\n")
     run_command_on_local(
         'sudo rm -rf ~/.ssh/known_hosts;'
@@ -1665,8 +1702,6 @@ def deploy_to_all(config):
         t.start()
     xen_master_node_q.join()
     safe_print("Finish step 3: assign ip to bond interfaces\n")
-    # wait long enough to let the IP assignment propagate
-    time.sleep(60)
 
     # step 4: change mgmt intf, using node_mgmtintf_q, on all run mgmtintf.sh
     for i in range(MAX_WORKERS):
@@ -1700,9 +1735,6 @@ def deploy_to_all(config):
     safe_print("CloudStack deployment finished\n")
 
 if __name__ == '__main__':
-    safe_print("Start to setup CloudStack for "
-               "Big Cloud Fabric %s\n" % (RELEASE_NAME))
-
     parser = argparse.ArgumentParser()
     parser.add_argument("-c", "--config-file", required=True,
                         help="CloudStack YAML config path")
@@ -1710,6 +1742,8 @@ if __name__ == '__main__':
     if not args.config_file:
         parser.error('--config-file is not specified.')
     else:
+        safe_print("Start to setup CloudStack for "
+                   "Big Cloud Fabric %s\n" % (RELEASE_NAME))
         config_file_path = args.config_file
         with open(config_file_path, 'r') as config_file:
             config = yaml.load(config_file)
